chore: ordering at parse
